pyproject.toml
[project]
name = "autoresearch-kaggle"
version = "0.1.0"
description = "Autonomous Kaggle competition research — agent-agnostic harness"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    # Core ML
    "numpy>=1.26",
    "pandas>=2.1",
    "scikit-learn>=1.5",
    "scipy>=1.12",

    # Gradient boosting
    "xgboost>=2.1",
    "lightgbm>=4.5",
    "catboost>=1.2",

    # Hyperparameter optimization
    "optuna>=4.0",

    # Deep learning (GPU)
    "torch>=2.4",
    "pytorch-tabnet>=4.1",

    # Feature engineering & encoding
    "category_encoders>=2.6",

    # AutoML (lightweight)
    "flaml>=2.3",

    # Analysis & visualization
    "matplotlib>=3.9",
    "seaborn>=0.13",

    # Utilities
    "pyarrow>=17.0",
    "joblib>=1.4",
    "imbalanced-learn>=0.12",

    # Scout layer
    "kaggle>=1.6",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
# Heavy AutoML - agent can install on demand
automl = [
    "autogluon.tabular>=1.2",
]
# Arxiv lookups for the Researcher phase (harness v2)
research = [
    "arxiv>=2.1",
]
# Observation dashboard (./am tui)
tui = [
    "textual>=0.80",
    "rich-pixels>=3.0",
    "textual-plotext>=1.0",
    "watchdog>=4.0",
    "pillow>=10",
]
# Optional GPU telemetry for the TUI (nvidia-smi fallback still used if omitted)
tui-gpu = [
    "pynvml>=11",
]
# Test runner
test = [
    "pytest>=8.0",
]

[project.scripts]
automedal-tui = "tui.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
